pyproject.toml
[project]
name = "repl-nix-shoppingspree"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.7.1",
    "pillow>=11.1.0",
    "pygame>=2.6.1",
    "streamlit>=1.41.1",
]
